refactor: page처리 중복되어 수정
home파일에 임시로 만든 page controller 삭제
