+++ src/main/java/com/coffice/app/branch/BranchController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.coffice.app.home.Pager;
+import com.coffice.app.page.Pager;
 
 import lombok.extern.slf4j.Slf4j;
