+++ src/main/java/com/coffice/app/branch/BranchService.java
@@ -5,7 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.coffice.app.home.Pager;
+import com.coffice.app.page.Pager;
+
 
 @Service
 public class BranchService {
