--- src/main/java/com/coffice/app/home/Pager.java
@@ -1,17 +1,0 @@
-package com.coffice.app.home;
-
-import lombok.Data;
-
-@Data
-public class Pager {
-	private String kind;
-	private String search;
-	
-	public String getSearch() {
-		if(this.search==null) {
-			this.search="";
-		}
-		return this.search;
-	}
-
-}
